Completed UI without Adjustment Feature
