+++ src/components/Inspect.jsx
@@ -1,27 +1,27 @@
-import { createSignal, createEffect } from "solid-js";
+import { createEffect } from "solid-js";
+import { Log } from "./Log";
 import '../styles/components/_inspect.scss';
 
 export const Inspect = (props) => {
-
-createEffect(() => {
-  props.record() ? 
+  createEffect(() => {
+    props.record() ? 
     document.querySelector("#recordButton > span").style.backgroundColor = "#2F2F33" : 
     document.querySelector("#recordButton > span").style.backgroundColor = "#D4D6D9";
-});
+  });
 
-const handleRecordClick = (e) => {
-  //when record button is triggered, all signals and application interaction is stored in cache;
-  e.preventDefault();
+  const handleRecordClick = (e) => {
+    //when record button is triggered, all signals and application interaction is stored in cache;
+    e.preventDefault();
 
-  //conditonal that checks if record is truthy then toggle back to falsly
-  props.setRecord(!props.record())
-  //need to update cache object to reflect the current state of the page
-};
+    //conditonal that checks if record is truthy then toggle back to falsly
+    props.setRecord(!props.record())
+    //need to update cache object to reflect the current state of the page
+  };
 
-const handleResetClick = (e) => {
-  console.log('Reset');
-  props.setCache(() => {});
-};
+  const handleResetClick = (e) => {
+    console.log('Reset');
+    props.setCache(() => {});
+  };
   //if falsy toggle to truthy and capture window data from browser
     //when reset button is triggered, cache storage is cleared 
         //if reset is truthy, reassing props.cache to empty cache
@@ -35,94 +35,9 @@ const handleResetClick = (e) => {
         <div id="resetButton" onClick={handleResetClick}>Reset</div>
       </div>
       <div id="history">
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">HI</div>
-        <div class="Update_Location">HI</div>
-        <div class="Update_Location">HI</div>
-        <div class="Update_Location">HI</div>
-        <div class="Update_Location">HI</div>
-        <div class="Update_Location">HI</div>
-        <div class="Update_Location">HI</div>
-        <div class="Update_Location">HI</div>
-        <div class="Update_Location">HI</div>
-        <div class="Update_Location">HI</div>
-        <div class="Update_Location">HI</div>
-        <div class="Update_Location">Hello</div>
-        <div class="Update_Location">Hello</div>
-        {/* <div id=""> </div> */}
-        {/* ^^^^ this is going contain the update location aspects that are updated by a loop pushing into an array*/}
+        <Log />
+        <Log />
+        <Log />
       </div>
     </div>
   );
