+++ src/components/Log.jsx
@@ -1,0 +1,16 @@
+import '../styles/components/_log.scss';
+
+export const Log = (props) => {
+
+  return(
+    <div class="logBox">
+      <div class="logTitle">Signal Title</div>
+      <div class="logContent">
+        <p>Updated computations only</p>
+        <p>Updated computations only</p>
+        <p>Updated computations only</p>
+        <p>Updated computations only</p>
+      </div>
+    </div>
+  )
+};
