--- src/styles/components/_buttons.scss
@@ -1,5 +1,0 @@
-#header{
-    float:left;
-    font-size:0.8rem;
-}
-
