+++ src/styles/components/_graph.scss
@@ -28,7 +28,6 @@
         // border:2px solid orange;
     }
 }
-
 .graphBox{
     position:relative;
     width:100%;
