+++ src/styles/components/_header.scss
@@ -11,11 +11,14 @@
 }
 
 #header {
+    position:absolute;
+    top:0;
     display:flex;
     justify-content:space-between;
     align-items:center;
     padding:0 m.$side-padding;
     width:100%; 
+    height:30px;
     background-color: m.$primary-color;
     cursor:pointer;
 
@@ -36,7 +39,7 @@
 
     &:hover{ opacity:0.35; }
     &.active{ opacity:1; }
-
+    
     &.hbox{
         &::before { margin-bottom:2px; }
         &::before, &::after {
