+++ src/styles/components/_inspect.scss
@@ -20,7 +20,6 @@ $icon-size: 22px;
     grid-template-rows: m.$box-height auto;
     border-right:1px solid m.$primary-color;
     overflow:hidden;
-    // border:8px solid orange;
 
     #logHead {
         display:flex;
@@ -56,7 +55,6 @@ $icon-size: 22px;
     #history { 
         width:100%; 
         height:100%; 
-        overflow: scroll;
-        // border:5px solid green;
+        overflow-y:scroll;
     }
 }
