+++ src/styles/components/_navbar.scss
@@ -1,6 +1,9 @@
 @use "../main.scss" as m;
 
 #navbar{
+    position:absolute;
+    bottom:0;
+    height: m.$box-height;
     display:flex;
     justify-content:space-around;
     align-items:center;
